Update createJob & DeleteJob

+++ Backend/controller/job.controller.js
@@ -2,6 +2,10 @@ import { Job } from "../model/job.model.js";
 
 export const createJob = async (req, res) => {
     try {
+        if (req.user.role !== "admin" && req.user.role !== "alumni") {
+            return res.status(403).json({ message: "Not authorized to post job" });
+        }
+        
         const job = await Job.create({
             ...req.body,
             postedBy: req.user.id
@@ -49,16 +53,25 @@ export const getJobById = async (req, res) => {
 
 export const updateJob = async (req, res) => {
     try {
-        const job = await Job.findByIdAndUpdate(
+    
+        const job = await Job.findById(req.params.id);
+        if (!job) {
+            return res.status(404).json({ message: "Job not found" });
+        }
+
+        if (req.user.role !== "admin" && job.postedBy.toString() !== req.user.id) {
+            return res.status(403).json({ message: "Not allowed to update this job" });
+        }
+
+        const updatedJob = await Job.findByIdAndUpdate(
             req.params.id,
             req.body,
             { new: true }
         );
-
         res.status(200).json({
             success: true,
             message: "Job updated successfully",
-            job
+            updatedJob
         });
     } catch (error) {
         res.status(500).json({ success: false, error: error.message });
@@ -68,8 +81,20 @@ export const updateJob = async (req, res) => {
 
 export const deleteJob = async (req, res) => {
     try {
+         const job = await Job.findById(req.params.id);
+        if (!job) {
+            return res.status(404).json({ message: "Job not found" });
+        }
+
+       
+        if (req.user.role !== "admin" && job.postedBy.toString() !== req.user.id) {
+            return res.status(403).json({ message: "Not allowed to delete this job" });
+        }
+
         await Job.findByIdAndDelete(req.params.id);
+
         res.status(200).json({ message: "Job deleted successfully" });
+
     } catch (error) {
         res.status(500).json({ error: error.message });
     }
